drawWorldMapHud/mapScreenDrawHud: (int)(u16) hint cast recovers signed cmpwi 0xad (92.9->93.1, 93.06->93.14)

+++ src/main/dll/dll_0000_gameui.c
@@ -1555,7 +1555,7 @@ void drawWorldMapHud(void)
         }
 
         lv = 0;
-        if ((u16)getNextTaskHintText() > 0xad)
+        if ((int)(u16)getNextTaskHintText() > 0xad)
         {
             lv = 1;
         }
@@ -2534,7 +2534,7 @@ void mapScreenDrawHud(int p1, int p2, int p3)
                 else li_ = -1;
             }
             lv = 0;
-            if ((u16)getNextTaskHintText() > 0xad)
+            if ((int)(u16)getNextTaskHintText() > 0xad)
             {
                 lv = 1;
             }
